requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
pillow
openpyxl
python-dateutil
psycopg2-binary
